requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-extensions==2.2.9
django-haystack==2.8.1
django-heroku==0.3.1
django-sphinxdoc==2.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.6.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.0.1
